requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
django-filter
django-mptt
drf-yasg